+++ src/plugin-api/main/uk/co/reecedunn/intellij/plugin/processor/intellij/log/QueryLogViewer.kt
@@ -15,6 +15,7 @@
  */
 package uk.co.reecedunn.intellij.plugin.processor.intellij.log
 
+import com.intellij.openapi.Disposable
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
@@ -22,7 +23,7 @@ import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.intellij.ui.content.ContentFactory
 
-class QueryLogViewer : ToolWindowFactory, DumbAware {
+class QueryLogViewer : ToolWindowFactory, DumbAware, Disposable {
     private var logView: QueryLogViewerUI? = null
 
     override fun createToolWindowContent(project: Project, toolWindow: ToolWindow) {
@@ -36,7 +37,12 @@ class QueryLogViewer : ToolWindowFactory, DumbAware {
         contentManager.addContent(content)
         contentManager.setSelectedContent(content)
 
-        Disposer.register(contentManager, logView!!)
+        Disposer.register(contentManager, this)
         toolWindow.show(null)
+    }
+
+    override fun dispose() {
+        logView?.let { Disposer.dispose(it) }
+        logView = null
     }
 }